+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.2.0
+* New Features implemented: 
+  * Set Bluetooth Connection method added to facilitate bluetooth connections between Device and Charging base.
+  * Start Scan method added.
+  * Stop Scan method added.
+
 ## 1.1.0+3
 
 * added scan product feature using peripherral app.